Font: add shortcut to the default keymap to add special characters
Adding special characters is not a very common operation, but it may
be useful for certain workflows to be able to do it using a shortcut.

The new shortcut is Ctrl + Shift + A in font edit mode: kind of like
Shift + A for Add, but that is of course already taken by "Insert A".