Anim: only show FCurve modifier toggle widget when there are modifiers
In the animation editors, only show the FCurve Modifier toggle widget
(the wrench) when the FCurve actually has one or more modifiers on it.

This simplifies the UI, by only showing the widget when it's relevant.

A possible downside is that now it's no longer possible to disable
modifiers before adding them.