GPU: Resolve runtime issues with texture views for Metal.
When creating a texture view, Metal may require that the original
GPUTexture state is modified in some way. This may be a result
of deferred creation, or, to cache the texture view against the
source.

As a result, GPUTexture passed into GPU_texture_create_view
cannot be const.

This PR also contains a number of texture usage flag fixes
for EEVEE Next for textures used as views.

Authored by Apple: Michael Parkin-White

Ref #96261