Check both retopology and xray for selection occlusion
There's no need to use selection occlusion when xray is enabled, since it doesn't work anyway (and is a nice way of disabling selection occlusion without disabling the retopology overlay).
Since it's used in multiple places I decided to make a define, though only for v3d, not one that takes overlay because xray lives in shading.