UI: replace "x" with multiplication sign when displaying calculation
The multiplication sign looks like an "x" but should be used in its
stead to display calculations and dimensions such as "1920x1080". It
is supported in many fonts including DejaVu Sans, the font currently
used for the UI.

The modification was done only for strings which are displayed in the
user interface, so this excludes prints.