Implement hack in config to compensate for purity lost due to Guard Rail LUT's precision
Using a hack in Guard Rail, use CDL sat to bring back a bit of purity, this causes wider gamut render to be a bit less smooth but the degree is acceptable. Might remove this hack in the future if better approach comes or when the community cares less about "looking exactly like display's native".