Fix missing assets in the source archive
The assets are required to build proper Blender release, so they
can not be skipped from packing.

The packing ignores the `working` directory as it seems to be big
and sounds that it is not needed for the release.

This change makes it so there is an extra directory inside the
archive, next to the sources directory. For example:

```
blender-3.6.0-alpha.tar.xz:
  - blender-3.6.0-alpha
  - lib
```

This matches the structure of the source archive with packages.