Implemented an explicit pairing function for simulation input/output.
This replaces the search logic in the simulation input node, which
previously set the `output_node_id` to pair an input to an output.
Now the pairing takes place right after nodes are created by the operator.

An operator will also be added that allows pairing simulation nodes in case
the pairing is lost somehow or one part of the input/output pair is deleted.
That shouldn't normally be happening, but it might be the result of python
scripting or some corner case. It's good to have a regular method of
correcting such errors.