Fix #106127: crash enabling "Dynamic" on already bound MeshDeform mod
The UI would already disable the prop, but setting via python or drivers
would still be possible, resulting in a crash because important dynamic
binding data would be missing in the modifier evaluation.

Now report an error in the modifier in such case instad of crashing.

Alternatively, re-binding under the hood could be investigated.