Modifiers: Boolean Fast Mode Refactor
This patch concentrates work on Boolean Fast Mode so that it can be used on operations whose geometries have coplanar faces (useful for T75913).

One thing to note is that, for simplification and efficiency, intersections between coplanar triangles are ignored. The intersections of the triangles connected to these coplanars are enough to solve the operation.

For coplanar triangles, a map grouping the faces is filled so that duplicate faces or inside coplanar faces are removed.

To Do:
- Profiling