UI: replace "clipboard" with "internal clipboard" when relevant
Following up the previous commit which replaces "[copy/paste] buffer"
with the newly-introduced "internal clipboard", this commit adds
"internal" before existing occurrences of "clipboard", when not
dealing with the system clipboard.

This distinction is useful so that the user will not expect something
copied to the clipboard (say a node or object) to be pastable to
another application.

I took the opportunity to slightly rephrase some descriptions closer
to the human interface guidelines, like using the imperative and
articles.

Review: https://projects.blender.org/blender/blender/pulls/106060