Nodes: make spacings unitless in String to Curves node
The String to Curves node allows selecting various metrics for the
characters. Size is the absolute height between two lines, and is
correctly marked as a distance. Character, Word and Line Spacing, on
the other hand, do not represent any distance, but are proportional to
the base size.

To make sure of this issue, try the following:
- using Metric as the unit system, create a String to Curves node,
  input a multi-line string into it and plug it into the output. Its
  size is 1m, as well as all its spacings.
- Now set its size to 2m. The spacings are still 1m, yet they have all
  doubled, in proportion to the basic size.
- Now set the unit system to Imperial and the length unit to Miles.
  Reset the size to 1 mi. The spacings are still at 1m = 0.000621 mi,
  which doesn't correspond to any measurement.
- Reset the Line Spacing to 1 mi and the second line of text is now
  many miles away.
- In fact a crash now occurs when trying to Frame Selected (numpad .).

This PR simply removes the properties' `PROP_DISTANCE` subtype.