Fix #106706: Fireflies with Nishita sky sun sampling at certain angles.
Fireflies may appear when using the Nishtia sky texture with
certain settings.

This appears to happen because the functions `pdf_uniform_cone()`
and `sky_radiance_nishita()` use slightly different mathematical
opertations to figure out if the ray is inside the sun disc or not.

The use of different functions results in loses in precision in
different places meaning the two functions do not agree on whether
or not some rays hit the sun disc or not.
Since they disagree, the PDF for when a ray samples the sky texture,
and ends up inside the sun disc may be incorrect,
resulting in fireflies.

This commit attempts to fix this issue by adjusting
`pdf_uniform_cone()` and `sky_radiance_nishita()` to use the same
math functions so they both agree on if a ray lies within the sun disc.