Geometry Nodes: Use implicit sharing in point/mesh conversion nodes
Reduces memory usage and time spent copying data in the "Points to
Vertices" and "Mesh to Points" nodes. An arbitrary speedup can be
observed by just adding more data to the source geometries.
For example, In a simple test I saw a 70x speedup. The sharing in
the "Mesh to Points" node is more limited though, since it can copy
from multiple domains.

A new utility function is added to share a custom data layer.
I'd imagine this being part of a more general "attribute interpolation"
abstraction in the future.