Shift angle adjustment code out of `shader.cpp` and into `shader_nodes.cpp`
As suggest by @brecht, we don't want the angle adjustment code in shader.cpp as that the Blender to Cycles converter, and so the code isn't accessible in the Hydra renderer or standalone Cycles.
Shifting it to shader_nodes.cpp allows the angle adjustment code to be accessible by Hydra and standalone Cycles.