Make first non-anonymous CustomData layer active
For each created CustomData layer check if the previous active layer
was an anonymous layer and if so set the new layer as active. Do the same
for the render/clone/mask layer. This way we never end up in a state
where a mesh has non-anonymous layers but none of them are active.