UI: Add Drag & Drop Feedback on Windows and x11
On Windows and x11, Drag & Drop data can be obtained when the
mouse enters in a window. This data can be used to provide feedback to
the user, for example, if the data contains a file path, it can be
displayed if an operation can be performed on the file on drop.