WindowManager: Create Context Based on Chosen Backend.
Issue found during development of the Vulkan backend. During
initialization there is always an OpenGL context being created.
This context can then be wrapped in a different backend what leads
to crashes.

This hasn't been detected when implementing the Metal backend. It
can be the case that OpenGL and Metal construct the same context
handle as OpenGL is a wrapper around Metal. Not sure so this needs
some proper testing.