Gpu/tests: Fix test_shader_compute_vbo
Made format variable non-static to avoid stack overflow
on second run of the test for backends other then OpenGl.
When format variable is static it will make Vulkan run
have 2 vertex attributes which will change the size of the
buffer needed to store the data on the device. This in turn
will lead to buffer overflow during GPU_vertbuf_read call.

Signed-off-by: Mateusz Albecki <mateusz.p.albecki@gmail.com>