WIP: VSE: Gradual retiming
This feature allows to smooth transition between 2 retimed segments.

Smooth transition is achieved by evaluating Y value of arc that is
tangent to lines of linear retimed segments. Because of this, when
transition length is changed both handles are moved symetrically. This
way it is possible to make transition without affecting content in
linear segments.

TODO:
 - Modal move operation results in precision issues. Not sure why, but I assume, that this is result of new `SEQ_retiming_remove_handle()` and workflow "delete gradient, and create new one".
 - It's probably best to make `SEQ_retiming_remove_handle()` work with gradients based on geometry. Workflow mentioned above also invalidates all pointers which makes code ugly.
 - Make visual feedback, that gradient will be added on shift key.
 - Fix label to show gradient.
 - Bunch of limits are probably missing everywhere.
 - Merge with sound support.