Geometry: Refactor creation to avoid creating empty data arrays
When creating a mesh with implicit sharing (or existing data), we often
don't want to allocate data only to replace it. The few existing cases
hack around this by creating empty geometry, setting the domain sizes,
and freeing the original builtin layers. This commit avoids the
creation of the empty data layers in the first place, so they don't
need to be freed. For example, there will be no ".edge_verts" attribute
on a mesh with only loose vertices.

Though we may want a more advanced API for creating geometry from
existing data eventually, this intermediate step does some useful
cleanup anyway, and makes creating geometries with shared builtin
layers less ugly, which is helpful for simulation cache reading.