Graph Editor: make selection account for individual handle status
This is a bit subtle.  Prior to this commit, all handles on a key
would be considered "selected" when determining what to select next,
even if only a single handle was selected.

For example, consider two keys, A and B, with main vertices very
close to each other, and with only the left bezier handle of A being
selected.  If the user then clicked on the main vertex of A, the
selection code would consider A already selected and thus select
B instead, despite the main vertex of A not being selected.  This
was a really weird behavior, and made it impossible to select the
main vertex of A in such situations.

This commit fixes that by properly considering the selection status
of each handle/vertex of a key individually.