Fix #107422: Mesh context menus do not open with the last used option
Caused by 99e5024e97f1c34697885b1a93935e72c019da5c

Above commit changed behavior in that it only flagged the `uiBlock` with
`UI_BLOCK_POPUP_MEMORY` in case a Menu's title was provided to
`ui_popup_menu_create_block`. Now when is the title provided? This
depends on the `bl_label` set in the menu class. But this is not always
the case, there are some menus that have a blank `bl_label` (as seen
with `VIEW3D_MT_edit_mesh_context_menu` for example -- this specifies its
own label inside the draw function depending on vertex/edge/face mode).
Thus the provided title is not really reliable as source information to
do the flagging (also see for example
a38b98478a61815f57dee29c226de61d9cd338e1), flagging should **always** be
done [and only check a title for actually inserting in the puphash].

Now why does this still work for menus **not** providing a title? This
is because `button_activate_exit` will also handle the puphash when
choosing a menu entry and does `ui_popup_menu_memory_set`. So we are all
good here.