Fix #107011: Boolean Modifier Overlap Threshold property issue in UI when value is 0
Suppport for non-linear slider was added in e95f71bf15, 0447aedb96.
For startvalue 0, value generated by logarithmic scale is 0 in
`ui_numedit_but_NUM()`. So the fix would to change softmin/hardmin (as
said in the commit message: 0447aedb96) of this property to its default value.

This fix also avoids the assert hit at `BLI_assert(value_step > 0.0f)`
which happens when value is 0 and clicked on side-arrows.