Fix: shading_ex_pie using bone X-ray when it does not apply
If an object is a deform child of an armature, and that armature is in
pose mode, and "Lock Object Modes" is disabled, and the child mesh is
selected, and "Extra Shading Pie Menu Items" is enabled, then the X-ray
toggle in the Shading Pie menu (Z) is incorrectly being applied for pose
bones.  Fix by adding missing selection logic to shading_ex_pie to
properly duplicate the logic of view32.toggle_xray.