UI: add Proportional Size to Proportional Edit popover
Rather than forcing the user to initiate a transform operation to edit the `Proportional Size`, allow editing of the `Proportional Size` through the UI.

---
cc623ee7b0fc implemented a change that didn't seem to have many implications.

The change was: do not save state changes in the scene related to Snap or Proportional Edit settings if the Transform operation is cancelled.

But this change didn't account for some users relies on the transform operation just to change the size of the Proportional Edit. The flow is:
- Press G or R or S to start a transform operation
- Scroll with the mouse to change the Proportional Size
- LMB or Esc to cancel the operation.

To avoid this inconsistent usage, this PR proposes to edit the Proportional Size value through the UI.

The affected headers are:
- View 3D
- Dop Sheet
- Image Editor
- Graph Editor
- Mask Editor